Move events ajax response to be outside of tickets commerce
